Move LightSensor functions to a seperate file
- Created a lightsensor.cpp and .hpp
- Moved all functions related to the light sensor to those files
